# src/main/resources/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Website API","description":"Управление сайтами"},{"name":"User API","description":"Управление пользователями"},{"name":"Topic API","description":"Управление топиками"},{"name":"Article API","description":"Управление статьями"}],"paths":{"/api/websites/user/{id}":{"get":{"tags":["Website API"],"summary":"Получить сайты по ID пользователя","operationId":"getUserWebsites","parameters":[{"name":"id","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Сайты найдены","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WebsiteList"}}}}}},"put":{"tags":["Website API"],"summary":"Обновить сайт","operationId":"updateWebsite","parameters":[{"name":"id","in":"path","description":"ID сайта","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebsiteData"}}},"required":true},"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Сайт обновлен","content":{"*/*":{"schema":{"type":"string"}}}}}},"patch":{"tags":["Website API"],"summary":"Обновить url сайта","operationId":"updateWebsiteUrl","parameters":[{"name":"id","in":"path","description":"ID сайта","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebsiteUrlData"}}},"required":true},"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Сайт обновлен","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/users/{id}":{"get":{"tags":["User API"],"summary":"Получение пользователя по id","operationId":"get","parameters":[{"name":"id","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Пользователь найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["User API"],"summary":"Обновление данных о пользователе","operationId":"update","parameters":[{"name":"id","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserData"}}},"required":true},"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Пользователь обновлен","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/websites/user":{"post":{"tags":["Website API"],"summary":"Создать сайт","operationId":"createWebsite","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebsiteData"}}},"required":true},"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"201":{"description":"Сайты создан","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Website"}}}}}}},"/api/users/register":{"post":{"tags":["User API"],"summary":"Регистрация нового пользователя","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserData"}}},"required":true},"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"201":{"description":"Пользователь создан","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/topics/create":{"post":{"tags":["Topic API"],"summary":"Создание топика","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TopicData"}}},"required":true},"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"201":{"description":"Топик создан","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Topic"}}}}}}},"/api/articles/user/{articleId}":{"patch":{"tags":["Article API"],"summary":"Обновление названия статьи","operationId":"updateArticleTitle","parameters":[{"name":"articleId","in":"path","description":"ID сайта","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleTitleData"}}},"required":true},"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Название статьи обновлено","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/websites/{id}":{"get":{"tags":["Website API"],"summary":"Получить сайт по ID","operationId":"get_1","parameters":[{"name":"id","in":"path","description":"ID сайта","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Сайт найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Website"}}}}}}},"/api/topics/{id}":{"get":{"tags":["Topic API"],"summary":"Получение топика по id","operationId":"get_2","parameters":[{"name":"id","in":"path","description":"ID топика","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Топик найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Topic"}}}}}}},"/api/topics/user/{userId}":{"get":{"tags":["Topic API"],"summary":"Получение топиков по ID пользователя","operationId":"getUserTopics","parameters":[{"name":"userId","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Топики найдены","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicList"}}}}}}},"/api/articles/user/{userId}":{"get":{"tags":["Article API"],"summary":"Получение статьи по id","operationId":"getUserArticles","parameters":[{"name":"userId","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Статья найдена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleListData"}}}}}}},"/api/websites/{userId}/{websiteId}":{"delete":{"tags":["Website API"],"summary":"Удалить сайт","operationId":"deleteWebsite","parameters":[{"name":"userId","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"websiteId","in":"path","description":"ID сайта","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Сайт удален","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/topics/{userId}/{topicId}":{"delete":{"tags":["Topic API"],"summary":"Удаление топика","operationId":"deleteTopic","parameters":[{"name":"userId","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"topicId","in":"path","description":"ID топика","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Статья не инициализована","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ArticleInitializationException"}}}},"404":{"description":"Топик не инициализован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicNotFoundException"}}}},"200":{"description":"Топик удален","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"ArticleInitializationException":{"type":"object","properties":{"cause":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"className":{"type":"string"},"nativeMethod":{"type":"boolean"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}},"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"className":{"type":"string"},"nativeMethod":{"type":"boolean"}}}},"message":{"type":"string"},"suppressed":{"type":"array","items":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"className":{"type":"string"},"nativeMethod":{"type":"boolean"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}}},"localizedMessage":{"type":"string"}}},"TopicNotFoundException":{"type":"object","properties":{"cause":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"className":{"type":"string"},"nativeMethod":{"type":"boolean"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}},"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"className":{"type":"string"},"nativeMethod":{"type":"boolean"}}}},"message":{"type":"string"},"suppressed":{"type":"array","items":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"className":{"type":"string"},"nativeMethod":{"type":"boolean"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}}},"localizedMessage":{"type":"string"}}},"WebsiteData":{"type":"object","properties":{"url":{"type":"string"},"description":{"type":"string"},"userId":{"type":"integer","format":"int64"}}},"UserData":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"username":{"type":"string"}}},"UserId":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"Website":{"type":"object","example":"Сущность сайта","properties":{"id":{"$ref":"#/components/schemas/WebsiteId","description":"ID","example":1},"url":{"type":"string","description":"url","example":"http://test.com"},"description":{"type":"string","description":"description","example":"Test description"},"creatorId":{"$ref":"#/components/schemas/UserId","description":"creator_id","example":1}}},"WebsiteId":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"User":{"type":"object","description":"Сущность пользователя","properties":{"id":{"$ref":"#/components/schemas/UserId","description":"ID","example":1},"email":{"type":"string","description":"E-mail","example":"test@mail.ru"},"password":{"type":"string","description":"password","example":"test1234"},"username":{"type":"string","description":"username","example":"test"}}},"TopicData":{"type":"object","properties":{"description":{"type":"string"},"userId":{"type":"integer","format":"int64"}}},"Topic":{"type":"object","description":"Сущность топика","properties":{"id":{"$ref":"#/components/schemas/TopicId","description":"ID","example":1},"description":{"type":"string","description":"description","example":"Test description"},"userId":{"$ref":"#/components/schemas/UserId","description":"user_id","example":1}}},"TopicId":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"WebsiteUrlData":{"type":"object","properties":{"url":{"type":"string"}}},"ArticleTitleData":{"type":"object","properties":{"title":{"type":"string"}}},"WebsiteList":{"type":"object","properties":{"websiteList":{"type":"array","items":{"$ref":"#/components/schemas/Website"}}}},"TopicList":{"type":"object","properties":{"topicList":{"type":"array","items":{"$ref":"#/components/schemas/Topic"}}}},"ArticleData":{"type":"object","properties":{"title":{"type":"string"},"url":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"topicDescription":{"type":"string"}}},"ArticleListData":{"type":"object","properties":{"articleDataList":{"type":"array","items":{"$ref":"#/components/schemas/ArticleData"}}}}}}}